% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedPPM.R
\name{expPerform_ingroup}
\alias{expPerform_ingroup}
\title{This is used for ingroup expected performance}
\usage{
expPerform_ingroup(data, LSL, USL, mean, sdIngroup)
}
\value{
ingroup expected performance
}
\description{
This is used for ingroup expected performance
}
