% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedPPM.R
\name{expPerform_overall}
\alias{expPerform_overall}
\title{This is used for overall expected performance}
\usage{
expPerform_overall(data, LSL, USL, mean, sdOver)
}
\value{
list of overall expected performance
}
\description{
This is used for overall expected performance
}
