\name{nir.spectra}
\alias{nir.spectra}
\docType{data}
\title{Near-infrared spectra of ternary mixtures}
\description{
  Data containing 95 near-infrared spectra of ternary
  mixtures of ethanol, water and iso-propanol. 
}
\usage{data(nir.spectra)}
\seealso{\code{\link[kohonen]{nir}}}
\references{F. Wulfert , W.Th. Kok, A.K. Smilde: Anal. Chem. 1998, 1761-1767}
\keyword{datasets}
