% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedPPM.R
\name{obPerform}
\alias{obPerform}
\title{This is used for observation performance}
\usage{
obPerform(data, LSL, USL, N)
}
\value{
list of observation performance
}
\description{
This is used for observation performance
}
