% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA-package.R
\docType{package}
\name{PCA-package}
\alias{PCA}
\alias{PCA-package}
\title{PCA: Principal Component Analysis Package}
\description{
The functions in this package will allow you to visualize principal component analysis (PCA) in either tabular form or graphically. Additionally, you will be able to determine whether the differences between these components are statistically significant.
}
\keyword{internal}
