% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundation.R
\name{pca.prepCanvas}
\alias{pca.prepCanvas}
\title{This is used to get prepare visualization}
\usage{
pca.prepCanvas(main = "Process Capability Analysis Graph",
  sub = "My PCA project", pca.col = c("#666666", "#BBBBBB", "#CCCCCC",
  "#DDDDDD", "#EEEEEE"))
}
\value{
prep Canvas
}
\description{
This is used to get prepare visualization
}
