% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca}
\alias{pca}
\title{Title}
\usage{
pca(X, standardize, threshold_var)
}
\arguments{
\item{X}{a data matrix X}

\item{standardize}{boolean declaring whether to standardize data}

\item{threshold_var}{the desired percentage of variance that needs to be explained by the principal components}
}
\value{
An object of class "principal components" containing transformed data, principal components and relevant information
}
\description{
Title
}
\examples{
data(iris)
pca_iris <- pca(iris[,-5], F, 0.95)
}
