% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pca_plot}
\alias{pca_plot}
\title{pca_plot}
\usage{
pca_plot(d, dataframe, groupby, frame = TRUE, frametype = "norm")
}
\arguments{
\item{d}{a numeric or complex data frame that provides the data for the analysis}

\item{dataframe}{the entire dataframe that d came from, that can have variables as factors to group the results}

\item{groupby}{the variable from dataframe that you want to color your plot by}

\item{frame}{a logical value indicating whether your groups should have a geometric shape drawn around them}

\item{frame.type}{the shape of the frame, can be 'norm', 't', or 'euclid'. Default is 'norm' which creates an ellipse around the groups}
}
\description{
Plots the results of a principal components analysis is 2 dimensions
}
\examples{
pca_plot(d = network_data[,10:130], dataframe = network_data, groupby = 'History3', frame = TRUE, frametype = 'norm')
}
\keyword{PCA}
\keyword{graph}
\keyword{plot}
