% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pca_tab}
\alias{pca_tab}
\title{pca_tab}
\usage{
pca_tab(x, retx = TRUE, center = TRUE, scale. = FALSE, tol = NULL, ...)
}
\arguments{
\item{x}{a numeric or complex data frame that provides the data for the analysis}

\item{retx}{a logical value indicating whether rotated variables should be returned}

\item{center}{a logical value indicating if variables should be shifted to be zero centered}

\item{scale.}{a logical value indicating whether variables should be scaled to have unit variance before analysis}

\item{tol}{a value indicating the magnitude below which components should be omitted}
}
\description{
Performs principal components analysis on a data matrix and returns the result in tabular form
}
\examples{
pca_tab(x = iris[1:4], scale. = TRUE)
}
\keyword{PCA}
\keyword{components}
