% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{PCAplot}
\alias{PCAplot}
\title{PCAplot for plotting Principal Components using ggplot2}
\usage{
PCAplot(gene, meta)
}
\arguments{
\item{gene}{gene data consists of gene names and samples of gene}

\item{meta}{a metadata explaining the hour of the day at which sample was taken}
}
\value{
a grid plot of two columns of which one is PCA plot of PC1 and PC2 and other is the hour of day when sample taken
}
\description{
PCAplot for plotting Principal Components using ggplot2
}
