\name{pcashinyapp-package}
\alias{pcashinyapp-package}
\alias{pcashinyapp}
\docType{package}
\title{
\packageTitle{pcashinyapp}
}
\description{
\packageDescription{pcashinyapp}
}



\references{
Package source is in the help file for package.skeleton
}

\keyword{ data }

\examples{
There are no examples
}