% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvalues.R
\name{pcvalues}
\alias{pcvalues}
\title{PC values}
\usage{
pcvalues(data = NULL, scale = FALSE, value = "null", digits = 4)
}
\arguments{
\item{data}{dataframe with quantitative variables}

\item{scale}{a boolean value to indicating whether the data should be scaled or not. Default \code{scale = FALSE}}

\item{value}{determines whether values outputted are "eigenvectors", "loadings", "eigenvalues", or "variance"}

\item{digits}{number of decimal digits. Default \code{digits = 4}}
}
\value{
a table with either eigenvectors, loadings, eigenvalues, or variance explained
}
\description{
\code{pcvalues()} produces results in a table format of eigenvectors, loadings, eigenvalues,
 and variance/accumulated proportion of variance explained of the principal components.
 The eigenvalues represent the amount of variance that can be explained by each principal component.
 The loadings represent the covariability of each variable with the principal component.
}
\examples{
pcvalues(data = pima_diabetes, scale = T, value = "eigenvectors")
pcvalues(data = pima_diabetes, scale = T, value = "loadings")
pcvalues(data = pima_diabetes, scale = T, value = "eigenvalues", digits = 3)
pcvalues(data = pima_diabetes, scale = T, value = "variance", digits = 3)
}
