% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot.principal_components}
\alias{plot.principal_components}
\title{Biplot for principal components}
\usage{
\method{plot}{principal_components}(x, ..., choices = c(1, 2), vectors = TRUE, vector_names = TRUE)
}
\arguments{
\item{x}{An object of class "principal_components"}

\item{...}{generic arguments}

\item{choices}{Indices declaring which of the PC's you want to plot}

\item{vectors}{A boolean declaring whether to include vectors representing the constituents of the principal components}

\item{vector_names}{A boolean declaring whether to label the vectors}
}
\value{
A biplot of the principal components equivalent to that described by Gower and Hand (1996)
}
\description{
Biplot for principal components
}
\examples{
pca_iris <- pca(iris[,-5], FALSE, 0.95)
plot(pca_iris, choices = c(1, 2),  vectors = TRUE, vector_names = TRUE)
}
