% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot2D}
\alias{plot2D}
\title{Plot 2D}
\usage{
plot2D(data = NULL, type = "original", arrows = FALSE)
}
\arguments{
\item{data}{dataframe with 2 variables}

\item{type}{scatterplot type, either original or principle component coordinates}

\item{arrows}{boolean variable. Indicates whether or not to plot arrows}
}
\description{
\code{plot2D()} produces a 2D scatter plot of the data, which can be customized via its arguments.
This function displays a scatter plot of the data with arguments to change orientation of the axes and overlay principle component vectors.
}
\examples{
pima_diabetes_2D <- scale(pima_diabetes[,1:2])
plot2D(data = pima_diabetes_2D)
plot2D(data = pima_diabetes_2D, type = "PC")
plot2D(data = pima_diabetes_2D, type = "PC", arrows = TRUE)
}
