% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{Plot 3D}
\usage{
plot3D(data = NULL, type = "original", arrows = FALSE)
}
\arguments{
\item{data}{dataframe with 3 variables}

\item{type}{scatterplot type, either original or principle component coordinates}

\item{arrows}{boolean variable. Indicates whether or not to plot arrows}
}
\description{
\code{plot3D()} produces an 3D interactive scatter plot of the data, which can be customized via its arguments.
 This function displays an interactive scatter plot of the data with arguments to change orientation of the axes
 and overlay principle component vectors.
}
\examples{
pima_diabetes_3D <- scale(pima_diabetes[,3:5])
plot3D(data = pima_diabetes_3D)
plot3D(data = pima_diabetes_3D, type = "PC")
plot3D(data = pima_diabetes_3D, type = "PC", arrows = TRUE)
}
