% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA_withoutScaling.R
\name{plotpca_withoutscaling}
\alias{plotpca_withoutscaling}
\title{Construct an interactive PCA plot (without data scaling) using the gene intensity data from different samples or individuals}
\usage{
plotpca_withoutscaling(genedata, metadata)
}
\arguments{
\item{genedata}{Gene_intensity_data metadata the_sample_metadata.}
}
\value{
The interactive PCA plot without scaling
}
\description{
This function make an interactive PCA plot using the gene intensity data for different samples or individuals. The PCA analysis is done without scaling the data.
}
