% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA_withScaling.R
\name{plotpca_withscaling}
\alias{plotpca_withscaling}
\title{Construct an interactive PCA plot (with data scaling) using the gene intensity data from different samples or individuals}
\usage{
plotpca_withscaling(genedata, metadata)
}
\arguments{
\item{genedata}{Gene_intensity_data metadata the_sample_metadata.}
}
\value{
The interactive PCA plot with scaling
}
\description{
This function make an interactive PCA plot using the gene intensity data for different samples or individuals. The PCA analysis is done after scaling the data.
}
