\name{pnts}
\alias{pnts5}
\alias{pnts2}
\docType{data}
\title{Artificial multivariate normal data sets}
\description{
  The pnts2 and pnts5 data sets are artificial multivariate data sets
  in two and five dimensions. Both contain 400 points and consist of the
  sum of two multivariate normal distributions, differing in the size of
  the diagonal elements of the covariance matrix by a factor of 32. This
  leads to distributions where many points are present very close to the
  origin.
}
\seealso{\code{\link{spheres}}}
\usage{data(pnts)}
\keyword{datasets}
