% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentialProcessCapability.R
\name{ppclist}
\alias{ppclist}
\title{This is used total Potential Process Capability in R}
\usage{
ppclist(data, LSL = NA, USL = NA, meanST = 0, sdTotal = 1)
}
\value{
potential process capability
}
\description{
This is used total Potential Process Capability in R
}
