% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{print.principal_components}
\alias{print.principal_components}
\title{Title}
\usage{
\method{print}{principal_components}(ob)
}
\arguments{
\item{ob}{An object of class "principal components"}
}
\value{
Prints basic information regarding the principal components
}
\description{
Title
}
\examples{
pca_iris <- pca(iris[,-5], FALSE, 0.95)
print(pca_iris)
}
