% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_this.R
\name{run_this}
\alias{run_this}
\title{This function perform PCA analysis
Please type "run_this()"in the console below}
\usage{
run_this()
}
\arguments{
\item{select}{your files from your documents}
}
\value{
PCA results of 7 variants

Anova tables of 7 variants

PCA graph
}
\description{
This function perform PCA analysis
Please type "run_this()"in the console below
}
