% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_func}
\alias{scale_func}
\title{Scale Variables for PCA}
\usage{
scale_func(x)
}
\arguments{
\item{x}{numeric variables used to determine principal components for subsequent PCA.}
}
\value{
The variable minus the mean of the variables, divided by standard deviation of variables.
}
\description{
This function allow us to scale the variable within a dataframe or matrix for PCA.
}
\examples{
scale_func(wbg)
}
\keyword{normalize}
\keyword{scale,}
\keyword{variables,}
