% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{screeplot.principal_components}
\alias{screeplot.principal_components}
\title{Screeplot for principal_components objectr}
\usage{
\method{screeplot}{principal_components}(x, ..., threshold = NULL)
}
\arguments{
\item{x}{An object of class "principal_components"}

\item{...}{generic arguments}

\item{threshold}{A threshold that you desire the components to cross, will be plotted if value given}
}
\value{
A screeplot showing the weighting and cumulative weighting of the components
}
\description{
Screeplot for principal_components objectr
}
\examples{
pca_iris <- pca(iris[,-5], FALSE, 0.95)
screeplot(pca_iris)
}
