% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SD_var}
\alias{SD_var}
\title{Standard deviation for variables in PCA}
\usage{
SD_var(x)
}
\arguments{
\item{x}{numeric variables used to determine principal components for subsequent PCA.}
}
\value{
the standard deviation of the variables, rounded to 2 significant digits.
}
\description{
This function allow us to find the standard deviations variable within columns of a dataframe or matrix for PCA.
}
\examples{
SD_var(wbg)
}
\keyword{dev,}
\keyword{deviation,}
\keyword{round,}
\keyword{standard}
\keyword{std}
