% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundation.R
\name{sdOverAll2}
\alias{sdOverAll2}
\title{This is used for getting total standard deviation.(when n ==1)}
\usage{
sdOverAll2(split, rowSplit, colSplit, nST, mean, na.rm = FALSE, c4 = 1)
}
\arguments{
\item{c(split, }{rowSplit, colSplit, nST, mean, na.rm = FALSE, c4 = 1)}
}
\value{
total standard devtiation for n==1
}
\description{
This is used for getting total standard deviation.(when n ==1)
}
