% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{summary.principal_components}
\alias{summary.principal_components}
\title{Summary for principal component object}
\usage{
\method{summary}{principal_components}(object, ...)
}
\arguments{
\item{object}{An object of class "principal_components"}

\item{...}{generic arguments}
}
\value{
A summary providing the weights of each of the components
}
\description{
Summary for principal component object
}
\examples{
pca_iris <- pca(iris[,-5], FALSE, 0.95)
summary(pca_iris)
}
