% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalProcessPerformance.R
\name{tpplist}
\alias{tpplist}
\title{This is used total Process Performance in R}
\usage{
tpplist(data, LSL = NA, USL = NA, meanST = 0, sdOver = 1)
}
\value{
total standard devtiation
}
\description{
This is used total Process Performance in R
}
