\name{winesLight}
\alias{winesLight}
\alias{winesLightA}
\alias{winesLightB}
\alias{winesLight.classes}
\alias{winesLightA.classes}
\docType{data}
\title{Wine data (subset)}
\description{
  Data set \code{winesLight} is a subset of the
  \code{\link[kohonen]{wines}} data 
  containing sixty wines and six measured variables; the variables
  included have numbers 2, 3, 6, 7, 9 and 12 in the original wines data
  set. Object \code{winesLight.classes} contains the class label.
  
  These data are the results of chemical analyses of
  wines grown in the same region in Italy but derived from three
  different cultivars. The data set contains the quantities of several
  constituents found in each of the three types of wines, as well as
  some spectroscopic information.

  Two other sets of wine samples are present with the same variables:
  \code{winesLightA} and \code{winesLightB}. For \code{winesLightA}
  samples, also class information is present (\code{winesLightA.classes}).
}
\seealso{\code{\link[kohonen]{wines}}}
\usage{data(winesLight)}
\source{\url{http://kdd.ics.uci.edu}}
\keyword{datasets}
