\name{yeast.alpha}
\alias{yeast.alpha}
\alias{yeast.alpha.classes}
\docType{data}
\title{Yeast data: alpha arrestation method}
\description{
  The yeast.alpha data contain the results of a microarray experiment in
  which the cell cycle of yeast has been studied. This particular subset
  contains data for 613 genes, measured at eighteen time points, for one
  particular arrestation method ('alpha'). A classification of the genes
  is available in variable \code{yeast.alpha.classes}.
}
\seealso{\code{\link[kohonen]{yeast}}}
\usage{data(yeast.alpha)}
\references{P. Spellman et al., Mol. Biol. Cell 9, 3273-3297 (1998)}
\keyword{datasets}
