% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{centre}
\alias{centre}
\title{Title}
\usage{
centre(X, standardize = F)
}
\arguments{
\item{X}{a matrix}

\item{standardize}{a boolean declaring whether to standardize}
}
\value{
X a centred matrix
}
\description{
Title
}
\examples{
X <- matrix(seq(1:9))
centre(X)
}
