% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA-package.R
\docType{package}
\name{PCA-package}
\alias{PCA}
\alias{PCA-package}
\title{PCA: This is a summary of what PCA is and how we can ran and visualize PCA in R}
\description{
Principal component analysis is a dimensionality-reduction method, that is used with multi-dimensional data sets, by transforming the variables into smaller components, without eliminating much of the data. Even though some of the accuracy may be compromised, PCA is great for simplifying very complicated and large data sets and exploring overall patterns as well as preparing the dataset for data visualization.
}
\keyword{internal}
