% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{print.principal_components}
\alias{print.principal_components}
\title{Print for principal component object}
\usage{
\method{print}{principal_components}(x, ...)
}
\arguments{
\item{x}{An object of class "principal_components"}

\item{...}{generic arguments}
}
\value{
Prints basic information regarding the principal components
}
\description{
Print for principal component object
}
\examples{
pca_iris <- pca(iris[,-5], FALSE, 0.95)
print(pca_iris)
}
