% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{summary.principal_components}
\alias{summary.principal_components}
\title{Title}
\usage{
\method{summary}{principal_components}(ob)
}
\arguments{
\item{ob}{An object of class "principal components"}
}
\value{
A summary providing the weights of each of the components
}
\description{
Title
}
\examples{
summary(pca_iris)
}
